authorization

+++ src/main/java/controller/actions/admin/contents/DeleteContentAction.java
@@ -3,6 +3,7 @@ package controller.actions.admin.contents;
 import controller.actions.IAction;
 import model.database.ContentMetaDAO;
 import model.database.DAO;
+import utilities.AccessControl;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.http.HttpServletRequest;
@@ -12,18 +13,24 @@ public class DeleteContentAction implements IAction {
     @Override
     public void doPost(HttpServletRequest request, HttpServletResponse response) throws Exception {
 
-        String doi = request.getParameter("doi");
+        if (AccessControl.isLoggedIn(request)) {
+            String doi = request.getParameter("doi");
 
-        DAO contentDAO = new ContentMetaDAO();
-        contentDAO.delete(doi);
-
-        response.sendRedirect("/admin/content");
+            DAO contentDAO = new ContentMetaDAO();
+            contentDAO.delete(doi);
+            response.sendRedirect("/admin/content");
+        } else {
+            response.sendRedirect("/admin");
+        }
     }
 
     @Override
     public void doGet(HttpServletRequest request, HttpServletResponse response) throws Exception {
-        RequestDispatcher requestDispatcher = request.getRequestDispatcher("jsp/admin/content/delete-content.jsp");
-        requestDispatcher.forward(request,response);
-
+        if (AccessControl.isLoggedIn(request)) {
+            RequestDispatcher requestDispatcher = request.getRequestDispatcher("jsp/admin/content/delete-content.jsp");
+            requestDispatcher.forward(request, response);
+        } else {
+            response.sendRedirect("/admin");
+        }
     }
 }
